tif_nfields should be of size_t type.

+++ libtiff/tiffiop.h
@@ -173,7 +173,7 @@ struct tiff {
 	TIFFPostMethod	tif_postdecode;	/* post decoding routine */
 /* tag support */
 	TIFFFieldInfo**	tif_fieldinfo;	/* sorted table of registered tags */
-	uint32		tif_nfields;	/* # entries in registered tag table */
+	size_t		tif_nfields;	/* # entries in registered tag table */
 	const TIFFFieldInfo *tif_foundfield;/* cached pointer to already found tag */
         TIFFTagMethods  tif_tagmethods; /* tag get/set/print routines */
         TIFFClientInfoLink *tif_clientinfo; /* extra client information. */
